Arreglado el fallo en la visualización de las primeras 4 muestras (no se inicializaban por un error en la algoritmia) Estandarizado el método de inicio de visualización para los 3 modos planteados. Arreglados algunos detalles relativos a buenas prácticas con Android. Externalizados Strings de elementos GUI.

+++ src/org/microlites/data/usb/UsbComThread.java
@@ -89,7 +89,7 @@ public class UsbComThread extends DataSourceThread {
 							//Apa??o para poder ver por pantalla algo coherente y comprobar si funcionaba bien
 							parser.step((byte) 0xda);
 							parser.step((byte) 0x00);
-							parser.step((byte) buffer.array()[i]);
+							parser.step(buffer.array()[i]);
 						}
 					}
 					buffer.clear();
